move the files from root into lib
