+++ packages/features/bookings/lib/getAllCredentialsForUsersOnEvent/refreshCredential.ts
@@ -8,16 +8,16 @@ import type { CredentialPayload } from "@calcom/types/Credential";
  * @param credential
  */
 export async function refreshCredential(credential: CredentialPayload): Promise<CredentialPayload> {
-    const newCredential = await prisma.credential.findUnique({
-      where: {
-        id: credential.id,
-      },
-      select: credentialForCalendarServiceSelect,
-    });
-  
-    if (!newCredential) {
-      return credential;
-    } else {
-      return newCredential;
-    }
+  const newCredential = await prisma.credential.findUnique({
+    where: {
+      id: credential.id,
+    },
+    select: credentialForCalendarServiceSelect,
+  });
+
+  if (!newCredential) {
+    return credential;
+  } else {
+    return newCredential;
   }
+}
