+++ packages/features/bookings/lib/getAllCredentialsForUsersOnEvent/refreshCredentials.ts
@@ -1,6 +1,8 @@
-import type { CredentialPayload } from "@calcom/types/Credential";
-import { refreshCredential } from "./refreshCredential";
 import async from "async";
+
+import type { CredentialPayload } from "@calcom/types/Credential";
+
+import { refreshCredential } from "./refreshCredential";
 
 /**
  * Refreshes the given set of credentials.
@@ -8,7 +10,7 @@ import async from "async";
  * @param credentials
  */
 export async function refreshCredentials(
-    credentials: Array<CredentialPayload>
-  ): Promise<Array<CredentialPayload>> {
-    return await async.mapLimit(credentials, 5, refreshCredential);
-  }
+  credentials: Array<CredentialPayload>
+): Promise<Array<CredentialPayload>> {
+  return await async.mapLimit(credentials, 5, refreshCredential);
+}
